Fix Ui issue and optimise UI
