+++ nsflow/frontend/src/components/trace/formatters.ts
@@ -1,0 +1,42 @@
+/*
+Copyright © 2025 Cognizant Technology Solutions Corp, www.cognizant.com.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// Shared number formatters for trace and analysis views.
+
+export type FormatCostOptions = {
+  zero?: "$0.00" | "—";
+};
+
+export const formatCost = (cost: number, opts: FormatCostOptions = {}): string => {
+  if (!Number.isFinite(cost) || cost <= 0) return opts.zero ?? "$0.00";
+  return `$${cost.toFixed(cost < 1 ? 4 : 2)}`;
+};
+
+export const formatTokens = (n: number): string => {
+  if (!Number.isFinite(n) || n <= 0) return "0";
+  if (n < 1000) return n.toString();
+  if (n < 1_000_000) return `${(n / 1000).toFixed(1)}K`;
+  return `${(n / 1_000_000).toFixed(2)}M`;
+};
+
+export const formatDuration = (seconds: number): string => {
+  if (!Number.isFinite(seconds) || seconds <= 0) return "0 ms";
+  if (seconds < 1) return `${Math.round(seconds * 1000)} ms`;
+  if (seconds < 60) return `${seconds.toFixed(2)} s`;
+  const m = Math.floor(seconds / 60);
+  const s = seconds - m * 60;
+  return `${m}m ${s.toFixed(1)}s`;
+};
