+++ nsflow/frontend/src/components/trace/traceConstants.ts
@@ -1,0 +1,33 @@
+/*
+Copyright © 2025 Cognizant Technology Solutions Corp, www.cognizant.com.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// Static config for the trace UI and its localStorage mirror.
+
+export const STORAGE_KEY = "nsflow:trace:invocations:v1";
+export const STORAGE_DEBOUNCE_MS = 200;
+export const UNATTACHED_INVOCATION_ID = "__unattached__";
+
+export const TREE_COL = {
+  chevron: 18,
+  bar: 140,
+  duration: 70,
+  pct: 60,
+  tokens: 70,
+  cost: 70,
+};
+
+export const GANTT_ROW_HEIGHT = 26;
+export const GANTT_LABEL_COL = 220;
