+++ nsflow/frontend/src/pages/Analysis/components/Section.tsx
@@ -1,0 +1,50 @@
+/*
+Copyright © 2025 Cognizant Technology Solutions Corp, www.cognizant.com.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import { ReactNode } from "react";
+import { Box, Typography, useTheme } from "@mui/material";
+
+// Titled block used to group each table on the Analysis page.
+const Section = ({
+  title,
+  subtitle,
+  children,
+}: {
+  title: string;
+  subtitle?: string;
+  children: ReactNode;
+}) => {
+  const theme = useTheme();
+  return (
+    <Box sx={{ mb: 4 }}>
+      <Box sx={{ display: "flex", alignItems: "center", justifyContent: "space-between", mb: 1 }}>
+        <Box>
+          <Typography variant="subtitle1" sx={{ fontWeight: 600, color: "text.primary" }}>
+            {title}
+          </Typography>
+          {subtitle && (
+            <Typography variant="caption" sx={{ color: theme.palette.text.secondary }}>
+              {subtitle}
+            </Typography>
+          )}
+        </Box>
+      </Box>
+      {children}
+    </Box>
+  );
+};
+
+export default Section;
