+++ nsflow/frontend/src/pages/Analysis/components/SummaryTile.tsx
@@ -1,0 +1,45 @@
+/*
+Copyright © 2025 Cognizant Technology Solutions Corp, www.cognizant.com.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import { Paper, Typography, alpha, useTheme } from "@mui/material";
+
+// One of the roll-up tiles at the top of the Analysis page.
+const SummaryTile = ({ label, value }: { label: string; value: string }) => {
+  const theme = useTheme();
+  return (
+    <Paper
+      variant="outlined"
+      sx={{
+        px: 2,
+        py: 1.5,
+        minWidth: 160,
+        backgroundColor: alpha(theme.palette.background.paper, 0.8),
+      }}
+    >
+      <Typography variant="caption" sx={{ color: theme.palette.text.secondary }}>
+        {label}
+      </Typography>
+      <Typography
+        variant="h6"
+        sx={{ fontWeight: 600, fontVariantNumeric: "tabular-nums", color: "text.primary" }}
+      >
+        {value}
+      </Typography>
+    </Paper>
+  );
+};
+
+export default SummaryTile;
